pyproject.toml
[tool.poetry]
name = "webserver"
version = "0.1.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
assistant = {path = "../Assistant/assistant", develop = true}
pylint = "^3.3.1"
boto3 = "^1.35.49"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
websockets = "^14.1"
python-socketio = {extras = ["asyncio-client"], version = "^5.11.4"}

sqlalchemy = "^2.0.36"
authlib = "^1.3.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.10"
itsdangerous = "^2.2.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "uvicorn:webserver.main:app --reload"
